Add files via upload
First draft, and the code I'll be submitting for my Mod_3 Homework Assignment.

Check it out and comment on it. I'd love any feedback about the use of the sleep() function or about the organization. In the future, I'll be be combining the different data processes into their own functions and it'll look much neater.